+++ CHANGELOG.md
@@ -1,6 +1,17 @@
+## 2.0.0
+
+* Update ZaloSDK v4.0
+* Add getAccessToken function
+* Update loginWithoutRefreshToken return oauthCode, codeVerifier
+
+## 1.0.0
+
+* Update version zalo 2.6
+
 ## 0.0.1
-Init project
-## 0.0.2
-Fix lint and update description
-## 0.0.3
-Fix example
+
+* Fork project from zalo_flutter
+* Init project
+* Fix lint and update description
+* Fix example
+